+++ src/ttyu.cc
@@ -8,8 +8,8 @@ int ttyu_ee_cb_call(ee__listener_t *l, v8::Local<v8::Value> data) {
   do {
     if(l->cb) {
       l->cb->Call(1, args);
+      ++count;
     }
-    ++count;
   } while((l = l->next));
   return count;
 }